refactor: refactored home page to use components
